+++ ExilenceNextApp/src/components/item-table/item-table-cell/ItemTableCell.styles.ts
@@ -2,7 +2,7 @@ import { makeStyles } from '@material-ui/core';
 
 const useStyles = makeStyles(theme => ({
   tableCell: {
-    padding: theme.spacing(0.75),
+    padding: theme.spacing(0.5),
     overflow: 'hidden',
     textOverflow: 'ellipsis',
     whiteSpace: 'nowrap'