+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [0.3.3] - NOT RELEASED
 ##### Added
 - Added a new column that displays the tab names an item exists in
+##### Changed
+- Reworked the stash tab dropdown entirely in the profile dialog
 ##### Fixed
 - Fixed a bug where some jewels and jewellry without value were listed in the item table
 - Fixed a bug where the support button would be placed on top of the scrollbar